Refactor MSBuild resolution
Copy the approach from `Microsoft.Build.Locator` to set environment variables for use when invoking the .NET CLI.
