+++ src/DotNetBumper.Core/Upgraders/DotNetCodeUpgrader.cs
@@ -52,7 +52,6 @@ internal sealed partial class DotNetCodeUpgrader(
             context.Status = StatusMessage($"Updating .NET code for {relativePath}...");
 
             (var fileResult, var fixes) = await ApplyFixesAsync(
-                upgrade.Channel,
                 fileName,
                 upgrade.SdkVersion,
                 cancellationToken);
@@ -78,47 +77,7 @@ internal sealed partial class DotNetCodeUpgrader(
         return result;
     }
 
-    private static Dictionary<string, string?> GetFormatEnvironment(Version channel, string sdkVersion)
-    {
-        var environment = new Dictionary<string, string?>()
-        {
-            ["NoWarn"] = "CA1515", // HACK Ignore CA1515 from .NET 9 as it seems to just break things
-        };
-
-        if (channel.Major > 7)
-        {
-            // HACK dotnet format seems to have issues resolving where the .NET SDK is installed
-            // with .NET 8. If MSBuildSDKsPath is set, it needs to be overridden otherwise it may
-            // point to an SDK version that is lower that the one we are upgrading to.
-            const string MSBuildSDKsPath = "MSBuildSDKsPath";
-
-            var dotnetRoot = Environment.GetEnvironmentVariable("DOTNET_ROOT");
-
-            if (string.IsNullOrEmpty(dotnetRoot))
-            {
-                dotnetRoot = Path.Combine(
-                    Environment.GetFolderPath(
-                        OperatingSystem.IsWindows() ?
-                        Environment.SpecialFolder.ProgramFiles :
-                        Environment.SpecialFolder.CommonApplicationData),
-                    "dotnet");
-            }
-
-            var msbuildSdksPath = Path.Combine(
-                dotnetRoot,
-                "sdk",
-                sdkVersion,
-                "Sdks");
-
-            // This has to be specifically set because DotNetProcess will otherwise unset it for other reasons
-            environment[MSBuildSDKsPath] = msbuildSdksPath;
-        }
-
-        return environment;
-    }
-
     private async Task<(ProcessingResult Result, Dictionary<string, int> Fixes)> ApplyFixesAsync(
-        Version channel,
         string projectOrSolution,
         NuGetVersion sdkVersion,
         CancellationToken cancellationToken)
@@ -138,7 +97,12 @@ internal sealed partial class DotNetCodeUpgrader(
             "--verbosity", Logger.GetMSBuildVerbosity(),
         ];
 
-        var environmentVariables = GetFormatEnvironment(channel, globalJson?.SdkVersion ?? sdkVersion.ToString());
+        var environmentVariables = new Dictionary<string, string?>()
+        {
+            ["NoWarn"] = "CA1515", // HACK Ignore CA1515 from .NET 9 as it seems to just break things
+        };
+
+        MSBuildHelper.TryAddSdkProperties(environmentVariables, globalJson?.SdkVersion ?? sdkVersion.ToString());
 
         var formatResult = await dotnet.RunAsync(
             workingDirectory,
